docs(Calendar): edit docs after review
make different files for examples, add some examples, add examples code, review links and structure
