+++ src/components/Calendar/__stories__/examples/CalendarExampleLocale/CalendarExampleLocale.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import frLocale from 'date-fns/locale/fr';
+
+import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
+import { StoryBookExample } from '../../../../../uiKit/components/StoryBookExample/StoryBookExample';
+import { Calendar } from '../../../Calendar';
+
+export type CalendarPropType = 'date' | 'date-range';
+
+export const CalendarExampleLocale = () => {
+  return (
+    <StoryBookExample className={cnDocsDecorator('Section')}>
+      <Calendar locale={frLocale} />
+    </StoryBookExample>
+  );
+};
