+++ src/components/Calendar/__stories__/examples/CalendarExampleView/CalendarExampleView.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import { addDays, startOfWeek } from 'date-fns';
-import frLocale from 'date-fns/locale/fr';
+import { startOfWeek } from 'date-fns';
 import ruLocale from 'date-fns/locale/ru';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
@@ -8,8 +7,6 @@ import { StoryBookExample } from '../../../../../uiKit/components/StoryBookExamp
 import { Calendar } from '../../../Calendar';
 
 export type CalendarPropType = 'date' | 'date-range';
-
-const events = [startOfWeek(new Date(), { locale: ruLocale }), new Date(), addDays(new Date(), 2)];
 
 export const CalendarExampleView = () => {
   return (
@@ -35,34 +32,10 @@ export const CalendarExampleViewSlider = () => {
   );
 };
 
-export const CalendarExampleDateMin = () => {
-  return (
-    <StoryBookExample className={cnDocsDecorator('Section')}>
-      <Calendar minDate={startOfWeek(new Date(), { locale: ruLocale })} maxDate={new Date()} />
-    </StoryBookExample>
-  );
-};
-
 export const CalendarExampleCurrent = () => {
   return (
     <StoryBookExample className={cnDocsDecorator('Section')}>
-      <Calendar currentVisibleDate={startOfWeek(new Date())} />
-    </StoryBookExample>
-  );
-};
-
-export const CalendarExampleEvents = () => {
-  return (
-    <StoryBookExample className={cnDocsDecorator('Section')}>
-      <Calendar events={events} />
-    </StoryBookExample>
-  );
-};
-
-export const CalendarExampleLocale = () => {
-  return (
-    <StoryBookExample className={cnDocsDecorator('Section')}>
-      <Calendar locale={frLocale} />
+      <Calendar currentVisibleDate={startOfWeek(new Date(), { locale: ruLocale })} />
     </StoryBookExample>
   );
 };
